+++ crates/orderbook/src/api/create_order.rs
@@ -2,6 +2,7 @@ use crate::orderbook::{AddOrderError, Orderbook};
 use anyhow::Result;
 use model::order::{OrderCreation, OrderUid};
 use shared::api::{error, extract_payload, internal_error, ApiReply, IntoWarpReply};
+use shared::order_validation::{PartialValidationError, ValidationError};
 use std::{convert::Infallible, sync::Arc};
 use warp::reply::with_status;
 use warp::{hyper::StatusCode, Filter, Rejection};
@@ -13,10 +14,170 @@ pub fn create_order_request() -> impl Filter<Extract = (OrderCreation,), Error =
         .and(extract_payload())
 }
 
+pub struct PartialValidationErrorWrapper(pub PartialValidationError);
+impl IntoWarpReply for PartialValidationErrorWrapper {
+    fn into_warp_reply(self) -> ApiReply {
+        match self.0 {
+            PartialValidationError::UnsupportedBuyTokenDestination(dest) => with_status(
+                error("UnsupportedBuyTokenDestination", format!("Type {dest:?}")),
+                StatusCode::BAD_REQUEST,
+            ),
+            PartialValidationError::UnsupportedSellTokenSource(src) => with_status(
+                error("UnsupportedSellTokenSource", format!("Type {src:?}")),
+                StatusCode::BAD_REQUEST,
+            ),
+            PartialValidationError::UnsupportedOrderType => with_status(
+                error(
+                    "UnsupportedOrderType",
+                    "This order type is currently not supported",
+                ),
+                StatusCode::BAD_REQUEST,
+            ),
+            PartialValidationError::Forbidden => with_status(
+                error("Forbidden", "Forbidden, your account is deny-listed"),
+                StatusCode::FORBIDDEN,
+            ),
+            PartialValidationError::InsufficientValidTo => with_status(
+                error(
+                    "InsufficientValidTo",
+                    "validTo is not far enough in the future",
+                ),
+                StatusCode::BAD_REQUEST,
+            ),
+            PartialValidationError::ExcessiveValidTo => with_status(
+                error("ExcessiveValidTo", "validTo is too far into the future"),
+                StatusCode::BAD_REQUEST,
+            ),
+            PartialValidationError::TransferEthToContract => with_status(
+                error(
+                    "TransferEthToContract",
+                    "Sending Ether to smart contract wallets is currently not supported",
+                ),
+                StatusCode::BAD_REQUEST,
+            ),
+            PartialValidationError::InvalidNativeSellToken => with_status(
+                error(
+                    "InvalidNativeSellToken",
+                    "The chain's native token (Ether/xDai) cannot be used as the sell token",
+                ),
+                StatusCode::BAD_REQUEST,
+            ),
+            PartialValidationError::SameBuyAndSellToken => with_status(
+                error(
+                    "SameBuyAndSellToken",
+                    "Buy token is the same as the sell token.",
+                ),
+                StatusCode::BAD_REQUEST,
+            ),
+            PartialValidationError::UnsupportedSignature => with_status(
+                error("UnsupportedSignature", "signing scheme is not supported"),
+                StatusCode::BAD_REQUEST,
+            ),
+            PartialValidationError::UnsupportedToken(token) => with_status(
+                error("UnsupportedToken", format!("Token address {token:?}")),
+                StatusCode::BAD_REQUEST,
+            ),
+            PartialValidationError::Other(err) => with_status(
+                internal_error(err.context("partial_validation")),
+                StatusCode::INTERNAL_SERVER_ERROR,
+            ),
+        }
+    }
+}
+
+pub struct ValidationErrorWrapper(ValidationError);
+impl IntoWarpReply for ValidationErrorWrapper {
+    fn into_warp_reply(self) -> ApiReply {
+        match self.0 {
+            ValidationError::Partial(pre) => PartialValidationErrorWrapper(pre).into_warp_reply(),
+            ValidationError::QuoteNotFound => with_status(
+                error(
+                    "QuoteNotFound",
+                    "could not find quote with the specified ID",
+                ),
+                StatusCode::BAD_REQUEST,
+            ),
+            ValidationError::InvalidQuote => with_status(
+                error(
+                    "InvalidQuote",
+                    "the quote with the specified ID does not match the order",
+                ),
+                StatusCode::BAD_REQUEST,
+            ),
+            ValidationError::PriceForQuote(err) => err.into_warp_reply(),
+            ValidationError::MissingFrom => with_status(
+                error(
+                    "MissingFrom",
+                    "From address must be specified for on-chain signature",
+                ),
+                StatusCode::BAD_REQUEST,
+            ),
+            ValidationError::WrongOwner(owner) => with_status(
+                error(
+                    "WrongOwner",
+                    format!("Address recovered from signature {owner} does not match from address"),
+                ),
+                StatusCode::BAD_REQUEST,
+            ),
+            ValidationError::InsufficientBalance => with_status(
+                error(
+                    "InsufficientBalance",
+                    "order owner must have funds worth at least x in his account",
+                ),
+                StatusCode::BAD_REQUEST,
+            ),
+            ValidationError::InsufficientAllowance => with_status(
+                error(
+                    "InsufficientAllowance",
+                    "order owner must give allowance to VaultRelayer",
+                ),
+                StatusCode::BAD_REQUEST,
+            ),
+            ValidationError::InvalidSignature => with_status(
+                error("InvalidSignature", "invalid signature"),
+                StatusCode::BAD_REQUEST,
+            ),
+            ValidationError::InsufficientFee => with_status(
+                error("InsufficientFee", "Order does not include sufficient fee"),
+                StatusCode::BAD_REQUEST,
+            ),
+            ValidationError::SellAmountOverflow => with_status(
+                error(
+                    "SellAmountOverflow",
+                    "Sell amount + fee amount must fit in U256",
+                ),
+                StatusCode::INTERNAL_SERVER_ERROR,
+            ),
+            ValidationError::TransferSimulationFailed => with_status(
+                error(
+                    "TransferSimulationFailed",
+                    "sell token cannot be transferred",
+                ),
+                StatusCode::BAD_REQUEST,
+            ),
+            ValidationError::ZeroAmount => with_status(
+                error("ZeroAmount", "Buy or sell amount is zero."),
+                StatusCode::BAD_REQUEST,
+            ),
+            ValidationError::IncompatibleSigningScheme => with_status(
+                error(
+                    "IncompatibleSigningScheme",
+                    "Signing scheme is not compatible with order placement method.",
+                ),
+                StatusCode::BAD_REQUEST,
+            ),
+            ValidationError::Other(err) => with_status(
+                internal_error(err.context("order_validation")),
+                StatusCode::INTERNAL_SERVER_ERROR,
+            ),
+        }
+    }
+}
+
 impl IntoWarpReply for AddOrderError {
     fn into_warp_reply(self) -> ApiReply {
         match self {
-            Self::OrderValidation(err) => err.into_warp_reply(),
+            Self::OrderValidation(err) => ValidationErrorWrapper(err).into_warp_reply(),
             Self::DuplicatedOrder => with_status(
                 error("DuplicatedOrder", "order already exists"),
                 StatusCode::BAD_REQUEST,
