+++ crates/orderbook/src/api/get_fee_and_quote.rs
@@ -1,3 +1,4 @@
+use crate::api::post_quote::OrderQuoteErrorWrapper;
 use anyhow::Result;
 use chrono::{DateTime, Utc};
 use ethcontract::{H160, U256};
@@ -126,7 +127,8 @@ pub fn get_fee_and_quote_sell(
                 quotes
                     .calculate_quote(&query.into())
                     .await
-                    .map(SellResponse::from),
+                    .map(SellResponse::from)
+                    .map_err(OrderQuoteErrorWrapper),
             ))
         }
     })
@@ -142,7 +144,8 @@ pub fn get_fee_and_quote_buy(
                 quotes
                     .calculate_quote(&query.into())
                     .await
-                    .map(BuyResponse::from),
+                    .map(BuyResponse::from)
+                    .map_err(OrderQuoteErrorWrapper),
             ))
         }
     })
