+++ crates/orderbook/src/api/post_quote.rs
@@ -1,8 +1,10 @@
+use super::create_order::PartialValidationErrorWrapper;
 use anyhow::Result;
 use model::quote::OrderQuoteRequest;
+use reqwest::StatusCode;
 use shared::{
-    api::{self, convert_json_response},
-    order_quoting::QuoteHandler,
+    api::{self, convert_json_response, rich_error, ApiReply, IntoWarpReply},
+    order_quoting::{CalculateQuoteError, OrderQuoteError, QuoteHandler},
 };
 use std::{convert::Infallible, sync::Arc};
 use warp::{Filter, Rejection};
@@ -19,7 +21,10 @@ pub fn post_quote(
     post_quote_request().and_then(move |request: OrderQuoteRequest| {
         let quotes = quotes.clone();
         async move {
-            let result = quotes.calculate_quote(&request).await;
+            let result = quotes
+                .calculate_quote(&request)
+                .await
+                .map_err(OrderQuoteErrorWrapper);
             if let Err(err) = &result {
                 tracing::warn!(?err, ?request, "post_quote error");
             }
@@ -28,17 +33,62 @@ pub fn post_quote(
     })
 }
 
+#[derive(Debug)]
+pub struct OrderQuoteErrorWrapper(pub OrderQuoteError);
+impl IntoWarpReply for OrderQuoteErrorWrapper {
+    fn into_warp_reply(self) -> ApiReply {
+        match self.0 {
+            OrderQuoteError::Validation(err) => {
+                PartialValidationErrorWrapper(err).into_warp_reply()
+            }
+            OrderQuoteError::CalculateQuote(err) => {
+                CalculateQuoteErrorWrapper(err).into_warp_reply()
+            }
+        }
+    }
+}
+
+pub struct CalculateQuoteErrorWrapper(CalculateQuoteError);
+impl IntoWarpReply for CalculateQuoteErrorWrapper {
+    fn into_warp_reply(self) -> ApiReply {
+        match self.0 {
+            CalculateQuoteError::Price(err) => err.into_warp_reply(),
+            CalculateQuoteError::SellAmountDoesNotCoverFee { fee_amount } => {
+                warp::reply::with_status(
+                    rich_error(
+                        "SellAmountDoesNotCoverFee",
+                        "The sell amount for the sell order is lower than the fee.",
+                        serde_json::json!({ "fee_amount": fee_amount }),
+                    ),
+                    StatusCode::BAD_REQUEST,
+                )
+            }
+            CalculateQuoteError::Other(err) => err.into_warp_reply(),
+        }
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
+    use anyhow::anyhow;
+    use chrono::{DateTime, NaiveDateTime, Utc};
     use ethcontract::{H160, U256};
     use model::{
         app_id::AppId,
         order::{BuyTokenDestination, SellTokenSource},
-        quote::{OrderQuoteSide, PriceQuality, QuoteSigningScheme, SellAmount, Validity},
+        quote::{
+            OrderQuote, OrderQuoteResponse, OrderQuoteSide, PriceQuality, QuoteSigningScheme,
+            SellAmount, Validity,
+        },
     };
+    use reqwest::StatusCode;
     use serde_json::json;
-    use warp::test::request;
+    use shared::{
+        api::response_body,
+        order_quoting::{CalculateQuoteError, OrderQuoteError},
+    };
+    use warp::{test::request, Reply};
 
     #[test]
     fn deserializes_sell_after_fees_quote_request() {
@@ -190,5 +240,54 @@ mod tests {
             .header("content-type", "application/json")
             .json(&request_payload);
         assert!(request.filter(&filter).await.is_err());
+    }
+
+    #[tokio::test]
+    async fn post_quote_response_ok() {
+        let quote = OrderQuote {
+            sell_token: Default::default(),
+            buy_token: Default::default(),
+            receiver: None,
+            sell_amount: Default::default(),
+            buy_amount: Default::default(),
+            valid_to: 0,
+            app_data: Default::default(),
+            fee_amount: Default::default(),
+            kind: Default::default(),
+            partially_fillable: false,
+            sell_token_balance: Default::default(),
+            buy_token_balance: Default::default(),
+        };
+        let order_quote_response = OrderQuoteResponse {
+            quote,
+            from: H160::zero(),
+            expiration: DateTime::<Utc>::from_utc(NaiveDateTime::from_timestamp(0, 0), Utc),
+            id: Some(0),
+        };
+        let response = convert_json_response::<OrderQuoteResponse, OrderQuoteErrorWrapper>(Ok(
+            order_quote_response.clone(),
+        ))
+        .into_response();
+        assert_eq!(response.status(), StatusCode::OK);
+        let body = response_body(response).await;
+        let body: serde_json::Value = serde_json::from_slice(body.as_slice()).unwrap();
+        let expected = serde_json::to_value(order_quote_response).unwrap();
+        assert_eq!(body, expected);
+    }
+
+    #[tokio::test]
+    async fn post_quote_response_err() {
+        let response = convert_json_response::<OrderQuoteResponse, OrderQuoteErrorWrapper>(Err(
+            OrderQuoteErrorWrapper(OrderQuoteError::CalculateQuote(CalculateQuoteError::Other(
+                anyhow!("Uh oh - error"),
+            ))),
+        ))
+        .into_response();
+        assert_eq!(response.status(), StatusCode::INTERNAL_SERVER_ERROR);
+        let body = response_body(response).await;
+        let body: serde_json::Value = serde_json::from_slice(body.as_slice()).unwrap();
+        let expected_error = json!({"errorType": "InternalServerError", "description": ""});
+        assert_eq!(body, expected_error);
+        // There are many other FeeAndQuoteErrors, but writing a test for each would follow the same pattern as this.
     }
 }
